# docs/swagger.yaml
basePath: /v1
definitions:
  models.APIError:
    properties:
      error_message:
        type: string
      error_status:
        type: integer
    type: object
...
  models.Image:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      hash:
        type: string
      id:
        type: integer
      iso_image:
        type: string
      path:
        type: string
      size:
        type: integer
      updated_at:
        type: string
    type: object
  models.ImageForm:
    properties:
      description:
        type: string
      hash:
        type: string
      iso_image:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
...
info:
  contact: {}
  description: VMware Imaging Appliances written in GO with full HTTP-REST API
  title: go-via
  version: "0.1"
paths:
...
  /images:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Image'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get all images
      tags:
      - images
    post:
      consumes:
      - application/json
      parameters:
      - description: Add image
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.ImageForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create a new image
      tags:
      - images
  /images/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Remove an existing image
      tags:
      - images
    get:
      consumes:
      - application/json
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get an existing image
      tags:
      - images
    patch:
      consumes:
      - application/json
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update an image
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.ImageForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update an existing image
      tags:
      - images
...
swagger: "2.0"